+++ CutTheRope/game/CTRGameObject.cs
@@ -21,12 +21,12 @@ namespace CutTheRope.game
                 int i = 100;
                 if (nSString2.characterAtIndex(0) == 'R')
                 {
-                    i = (int)((float)(int)RTD((double)nSString2.substringFromIndex(2).intValue()) * 3.3f) / 2 + 1;
+                    i = (int)((int)RTD(nSString2.substringFromIndex(2).intValue()) * 3.3f) / 2 + 1;
                 }
                 float m_ = xml["moveSpeed"].floatValue() * 3.3f;
                 float r_ = xml["rotateSpeed"].floatValue();
                 CTRMover cTRMover = (CTRMover)new CTRMover().initWithPathCapacityMoveSpeedRotateSpeed(i, m_, r_);
-                cTRMover.angle_ = (double)rotation;
+                cTRMover.angle_ = rotation;
                 cTRMover.angle_initial = cTRMover.angle_;
                 cTRMover.setPathFromStringandStart(nSString2, vect(x, y));
                 setMover(cTRMover);
