+++ CutTheRope/game/Sock.cs
@@ -48,7 +48,7 @@ namespace CutTheRope.game
             b1.x = t1.x;
             b2.x = t2.x;
             b1.y = b2.y = y + 15f;
-            angle = (double)DEGREES_TO_RADIANS(rotation);
+            angle = DEGREES_TO_RADIANS(rotation);
             t1 = vectRotateAround(t1, angle, x, y);
             t2 = vectRotateAround(t2, angle, x, y);
             b1 = vectRotateAround(b1, angle, x, y);
